Terms moved from overview to splash
