+++ app/src/main/java/io/gnosis/safe/ui/splash/SplashActivity.kt
@@ -2,13 +2,14 @@ package io.gnosis.safe.ui.splash
 
 import android.content.Intent
 import android.os.Bundle
-import android.os.Handler
 import io.gnosis.safe.HeimdallApplication
 import io.gnosis.safe.databinding.ActivitySplashBinding
 import io.gnosis.safe.di.components.DaggerViewComponent
 import io.gnosis.safe.di.modules.ViewModule
-import io.gnosis.safe.ui.base.BaseActivity
 import io.gnosis.safe.ui.StartActivity
+import io.gnosis.safe.ui.base.BaseActivity
+import io.gnosis.safe.ui.safe.terms.TermsBottomSheetDialog
+import kotlinx.android.synthetic.main.activity_splash.*
 import javax.inject.Inject
 
 class SplashActivity : BaseActivity() {
@@ -22,9 +23,12 @@ class SplashActivity : BaseActivity() {
         super.onCreate(savedInstanceState)
         setContentView(binding.root)
         inject()
-        Handler().postDelayed({
-            startActivity(Intent(this, StartActivity::class.java))
-        }, 500)
+        viewModel.termsBottomSheetDialog = TermsBottomSheetDialog(this)
+        continue_button.setOnClickListener {
+            viewModel.checkTerms {
+                startActivity(Intent(this, StartActivity::class.java))
+            }
+        }
     }
 
     private fun inject() {
